make chunk coords of int type

+++ core/src/main/java/io/github/labyrinthgenerator/pages/game3d/chunks/Chunk.java
@@ -6,13 +6,13 @@ import io.github.labyrinthgenerator.pages.game3d.constants.Constants;
 import java.util.Objects;
 
 public class Chunk {
-    public final float x;
-    public final float z;
+    public final int x;
+    public final int z;
     public final int width = Constants.CHUNK_SIZE;
     public final int depth = Constants.CHUNK_SIZE;
     public final Vector2 center;
 
-    public Chunk(float x, float z) {
+    public Chunk(int x, int z) {
         this.x = x;
         this.z = z;
         center = new Vector2(x + width / 2f, z + depth / 2f);
